feat: update cache

+++ backend/springboot/baseweb/src/main/java/com/hust/baseweb/applications/programmingcontest/entity/ContestEntity.java
@@ -173,7 +173,7 @@ public class ContestEntity {
     private String evaluateBothPublicPrivateTestcase;
 
     @Column(name="min_time_between_two_submissions")
-    private int minTimeBetweenTwoSubmissions;
+    private long minTimeBetweenTwoSubmissions;
 
     @Column(name="judge_mode")
     private String judgeMode; // synchronous or asynchronous using queue
