+++ backend/springboot/baseweb/src/main/java/com/hust/baseweb/applications/programmingcontest/model/ModelGetContestDetailResponse.java
@@ -24,7 +24,7 @@ public class ModelGetContestDetailResponse {
     private String useCacheContestProblem;
     private String evaluateBothPublicPrivateTestcase;
     private int maxSourceCodeLength;
-    private int minTimeBetweenTwoSubmissions;
+    private long minTimeBetweenTwoSubmissions;
     private String judgeMode;
     private List<String> listStatusIds;
     private List<String> listSubmissionActionTypes;
